Add technique selection
Techniques assigned to shader groups are rendered onto the the shader group's render target if possible, otherwise the next shader group encountered with this technique assigned or a shader group with no assigned techniques. If no empty group exists and no shader group assigned with this technique is encountered, it is not rendered at all.